pyproject.toml
[tool.poetry]
name = "skill-stake-backend"
version = "0.1.0"
description = ""
authors = ["yaswanth <yaswanthsaivendra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
python-jose = "^3.3.0"
python-dotenv = "^1.0.1"
clerk-backend-api = "^1.6.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
